+++ reactivestate-compose/src/composeMain/kotlin/com/ensody/reactivestate/compose/ViewModelExt.kt
@@ -13,13 +13,12 @@ import androidx.lifecycle.viewmodel.compose.viewModel
 import com.ensody.reactivestate.ContextualErrorsFlow
 import com.ensody.reactivestate.ContextualStateFlowStore
 import com.ensody.reactivestate.ContextualValRoot
-import com.ensody.reactivestate.CoroutineLauncher
 import com.ensody.reactivestate.DI
 import com.ensody.reactivestate.ExperimentalReactiveStateApi
 import com.ensody.reactivestate.InMemoryStateFlowStore
 import com.ensody.reactivestate.ReactiveStateContext
+import com.ensody.reactivestate.ReactiveViewModel
 import com.ensody.reactivestate.invokeOnCompletion
-import com.ensody.reactivestate.triggerOnInit
 import com.ensody.reactivestate.withSpinLock
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.MainScope
@@ -39,14 +38,14 @@ import kotlinx.coroutines.sync.Mutex
  */
 @ExperimentalReactiveStateApi
 @Composable
-public inline fun <reified VM : CoroutineLauncher> reactiveViewModel(
+public inline fun <reified VM : ReactiveViewModel> reactiveViewModel(
     key: String? = null,
     crossinline onError: (Throwable) -> Unit,
     crossinline provider: ReactiveStateContext.() -> VM,
 ): State<VM> =
     onViewModel(key = key) {
         provider().also {
-            it.triggerOnInit()
+            it.onInit.trigger()
         }
     }.also { viewModel ->
         LaunchedEffect(viewModel.value) {