+++ reactivestate-core/src/commonMain/kotlin/com/ensody/reactivestate/ReactiveViewModel.kt
@@ -10,11 +10,12 @@ import kotlinx.coroutines.flow.MutableStateFlow
  *
  * The [scope] is meant to be filled with [ContextualVal] entries.
  *
- * For example, the [ContextualOnInit] hook can be used to launch coroutines via a [CoroutineLauncher]
- * directly from the ViewModel constructor, so you can have nicer error handling and loading indicators.
+ * The [onInit] hook can be used to launch coroutines via a [CoroutineLauncher] directly from the ViewModel constructor,
+ * so you can have nicer error handling and loading indicators.
  */
 @ExperimentalReactiveStateApi
-public abstract class ReactiveViewModel(final override val scope: CoroutineScope) : CoroutineLauncher {
+public abstract class ReactiveViewModel(public final override val scope: CoroutineScope) : CoroutineLauncher {
+    public val onInit: OnInit = ContextualStore.get(scope).getOrPut(OnInitKey) { OnInit(this) }
     private val emittedErrors: MutableFlow<Throwable> = ContextualErrorsFlow.get(scope)
     override val loading: MutableStateFlow<Int> = ContextualLoading.get(scope)
     public val stateFlowStore: StateFlowStore by lazy { ContextualStateFlowStore.get(scope) }
@@ -23,6 +24,8 @@ public abstract class ReactiveViewModel(final override val scope: CoroutineScope
         emittedErrors.tryEmit(error)
     }
 }
+
+private val OnInitKey = ContextualValStore.Key<OnInit>()
 
 @ExperimentalReactiveStateApi
 public val ContextualErrorsFlow: ContextualVal<MutableFlow<Throwable>> = ContextualVal("ContextualErrorsFlow") {