+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added reactive dependency injection solution (see `DI`) with graph tracking that can even re-create ViewModels when a DI module/node is changed at runtime.
 * Added `ContextualVal`/`ContextualValSuspend` for coroutine-local variables (like thread-locals or `CompositionLocal` for coroutines).
 * Made the experimental `ReactiveViewModel` more flexible by building it around `ContextualVal`.
+* Added experimental `OnInit` API for initial loading of ViewModel data with error and retry handling.
 * Added Composable `rememberOnViewModel { ... }` which remembers the block's result across configuration changes.
 * Added Composable `DI.derivedState { ... }` and `DI.derivedValue { ... }` which remember the block's result until a dependency changes and re-render the UI when needed.
 * Changed `loading` and all `withLoading` parameters from `MutableValueFlow` to `MutableStateFlow`.